Switch to switch statement LOL
When we have a bunch of if statements, we generally want to use a switch
statement. In the past, we couldn't switch based on a type. It had to be
a constant expression. Starting with C# 7.0 though, we can employ simple
pattern matching.

So let's convert that big if statement into a switch statement.
In my case, I let my IDE do it.

See how much cleaner that is!