Now fix it so the query compiles
At first glance, it seems that `user.Comments` can be null. It's
true if we're executing the code. But we're creating an expression
here that is not executed, but translated into a SQL query by EF Core.

Hence we don't need to add a null check and can use the `!` here to
let C# know, it's fiiiiine.