Now implement that same method on `Issue`
Suppose we want to "override" the extensio method. We can implement the
same method on `Issue` and it'll take precedence when run against an
`Issue`.

Notice how the last assert fails. This shows a limitation of extension
methods.

Extension methods are resolved at compile time, based on the declared
type of the variable. So even though the runtime type of `entity` is
`Issue`, the declared type is `IEntity` so the extension method is
invoked, even though `Issue` "overrode" the implementation.

This is a limitation because it means implementations can't provide
a better implementation for an extension method.