Enable nullable ref types for one file
Here we'll enable nullable reference types for a single class and
fixup that class.

There are a two main ways we could have fixed the compilation error.

1. Add a constructor that takes in a `string` and assign `Title` to
that `string`.
2. Initialize the property to an empty string (whether in a ctor or
as a property initializer).

In this case, I plan to use EF Core to hydrate `Issue` from the
database. Since we're not the ones in charge of instantiating issue,
we can set the property to `null!` which means it is ok for it to
be null when constructed, but we expect it won't be null by the time
someone references it because whatever is responsible for creating it
will populate that property.